requirements.txt
amqp==5.3.1
asgiref==3.9.0
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
eventlet==0.40.1
greenlet==3.2.3
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.2.0
reportlab==4.4.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
